pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "google-generativeai>=0.8.5",
    "pydrive2>=1.21.3",
    "python-dotenv>=1.2.1",
    "telegram>=0.0.1",
]
